requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
opencv-python>=4.8.1.78
Pillow>=10.2.0
imagehash>=4.3.1
python-whois>=0.8.0
aiofiles>=23.2.1
pydantic>=2.5.0
beautifulsoup4>=4.12.2
requests>=2.31.0
lxml>=4.9.3
python-dotenv>=1.0.0
